Frontend Update
Added the articles to the application and the ability to navigate to them once pressing on the image or title. Put the buttons on the bill screen and changed the styling on the Home Screen to accommodate newer iPhone models. Added more assets and a folder for the article screens.